增加返回分页数据

+++ response/response_paging.go
@@ -3,6 +3,6 @@ package response
 type RespPaging struct {
 	RespBody
 	Total    int64 `json:"total"`
-	Current  int64 `json:"current"`
-	PageSize int64 `json:"pageSize"`
+	Current  int   `json:"current"`
+	PageSize int   `json:"pageSize"`
 }
